refactor(metadata_allocator): split into module & cache one empty region;
Extract RegionHeader into region.rs submodule. Add single-region cache
to avoid mmap/munmap syscall churn on repeated alloc-dealloc cycles.
Simplify dealloc to skip intermediate list insertion when transitioning
from full directly to empty. Remove unnecessary unsafe from new().
